Relocate user anonymization helper
