+++ code/server/helpers.py
@@ -15,6 +15,7 @@ import json
 import logging
 import time
 import random
+import hashlib
 from discord.ext import commands
 from datetime import datetime, timezone
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple
@@ -1392,3 +1393,49 @@ def _safe_mid(msg: dict) -> int | None:
         return mid if mid > 0 else None
     except Exception:
         return None
+    
+# =============================================================================
+# User Anonymization
+# =============================================================================
+_ANON_ADJECTIVES = [
+    "Swift", "Brave", "Clever", "Gentle", "Fierce", "Quiet", "Bold", "Calm",
+    "Eager", "Happy", "Jolly", "Kind", "Lucky", "Merry", "Noble", "Proud",
+    "Quick", "Sharp", "Wise", "Zesty", "Amber", "Azure", "Coral", "Crimson",
+    "Golden", "Ivory", "Jade", "Misty", "Rusty", "Silver", "Sunny", "Velvet",
+    "Wild", "Cosmic", "Dreamy", "Frosty", "Glowing", "Hidden", "Lunar", "Mystic",
+    "Nimble", "Radiant", "Shadow", "Stormy", "Thunder", "Vivid", "Wandering", "Zen",
+]
+
+_ANON_NOUNS = [
+    "Fox", "Wolf", "Bear", "Hawk", "Owl", "Deer", "Lynx", "Crow",
+    "Otter", "Raven", "Tiger", "Eagle", "Falcon", "Heron", "Panda", "Viper",
+    "Dragon", "Phoenix", "Griffin", "Sphinx", "Kraken", "Wyrm", "Hydra", "Titan",
+    "Star", "Moon", "Comet", "Nova", "Nebula", "Quasar", "Void", "Storm",
+    "River", "Mountain", "Forest", "Ocean", "Glacier", "Canyon", "Meadow", "Valley",
+    "Knight", "Mage", "Rogue", "Sage", "Scout", "Warden", "Hunter", "Seeker",
+]
+
+
+def _anonymize_user(user_id: int | str) -> tuple[str, str]:
+    """
+    Generate a deterministic anonymous identity for a user.
+
+    Returns:
+        tuple of (anonymized_name, avatar_url)
+        - Name format: [Adjective][Noun][0-999]
+        - Avatar: Random image from picsum.photos (seeded by user_id)
+    """
+    uid = int(user_id) if user_id else 0
+    h = hashlib.sha256(str(uid).encode()).hexdigest()
+
+    adj_idx = int(h[:8], 16) % len(_ANON_ADJECTIVES)
+    noun_idx = int(h[8:16], 16) % len(_ANON_NOUNS)
+    num = int(h[16:24], 16) % 1000
+
+    name = f"{_ANON_ADJECTIVES[adj_idx]}{_ANON_NOUNS[noun_idx]}{num}"
+
+    # Use a seed for picsum to get consistent image per user
+    avatar_seed = int(h[24:32], 16) % 1000000
+    avatar_url = f"https://picsum.photos/seed/{avatar_seed}/200"
+
+    return name, avatar_url
